Separate some more functions from main.c
