+++ src/include/ascii/ascii.h
@@ -1,0 +1,12 @@
+#ifndef ASCII_H_
+#define ASCII_H_
+
+/* from https://www.asciiart.eu/animals/cats */
+static char *ascii_default = " |\\'/-..--.\n"
+                             " / _ _   ,  ;\n"
+                             "`~=`Y'~_<._./\n"
+                             " <`-....__.'";
+
+
+
+#endif
