+++ src/include/session/session.h
@@ -1,0 +1,6 @@
+#ifndef SESSION_H_
+#define SESSION_H_
+
+char *get_desktop_session();
+
+#endif
